requirements.txt
keras==2.12.0
matplotlib==3.6.3
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
scikit_learn==1.2.2
tensorflow==2.12.0
